review: strategy 패턴을 사용해 변경
