+++ src/main/kotlin/calculator/DelimiterStrategy.kt
@@ -1,0 +1,5 @@
+package calculator
+
+interface DelimiterStrategy {
+    fun parse(input: String): List<Int>?
+}
