+++ src/main/kotlin/calculator/ExpressionParser.kt
@@ -2,30 +2,13 @@ package calculator
 
 object ExpressionParser {
 
-    private const val DEFAULT_DELIMITERS = "[,:]"
-    private const val CUSTOM_DELIMITER = "//(.)\n(.*)"
+    fun hasCustomDelimiter(input: String): Boolean =
+        input.startsWith("//") && input.contains("\n")
 
-    fun parse(input: String): List<Int> {
-        Regex(CUSTOM_DELIMITER).find(input)?.let { result ->
-            return parseByCustomDelimiter(result)
-        }
-
-        Regex(DEFAULT_DELIMITERS).find(input)?.let { _ ->
-            return parseByDefaultDelimiter(input)
-        }
-
-        return listOf(parseUniNumber(input))
-    }
-
-    private fun parseByCustomDelimiter(result: MatchResult): List<Int> {
-        val customDelimiter = result.groupValues[1]
-        val tokens = result.groupValues[2].split(customDelimiter)
-        return tokens.map { token -> token.toInt() }
-    }
-
-    private fun parseByDefaultDelimiter(input: String): List<Int> =
-        input.split(DEFAULT_DELIMITERS.toRegex())
-            .map { it.toInt() }
+    fun parse(
+        input: String,
+        delimiterStrategy: DelimiterStrategy
+    ): List<Int> = delimiterStrategy.parse(input) ?: listOf(parseUniNumber(input))
 
     private fun parseUniNumber(input: String): Int {
         val uniNumber = input.toIntOrNull()
