+++ src/main/kotlin/calculator/StringCalculator.kt
@@ -4,6 +4,14 @@ class StringCalculator {
 
     fun calculate(input: String?): Int = when {
         input.isNullOrEmpty() -> 0
-        else -> ExpressionParser.parse(input).sumOf { it }
+        else -> {
+            val strategy = if (ExpressionParser.hasCustomDelimiter(input)) {
+                CustomDelimiterStrategy()
+            } else {
+                DefaultDelimiterStrategy()
+            }
+
+            ExpressionParser.parse(input, strategy).sumOf { it }
+        }
     }
 }
